Fixes of p2p download

+++ apl-api/src/main/java/com/apollocurrency/aplwallet/api/p2p/FileChunkInfo.java
@@ -11,9 +11,9 @@ import java.util.Objects;
  */
 public class FileChunkInfo {
     public String fileId;
-    public int chunkId;
-    public long offset;
-    public int size;
+    public Integer chunkId;
+    public Long offset;
+    public Long size;
     /**
      * 0 - not present
      * 1 - download in progress;
